old/vqa-backend/requirements.txt
Flask==2.0.3
psycopg2-binary==2.9.3
torch==1.10.2
transformers==4.12.3
Pillow==8.4.0